+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 * The precompiler now works when used on a project targeting a framework earlier than .NET Core 3.1.
 * The `Position.Advance` method got a new overload that accepts a `ReadOnlySpan` of characters. It is recommended over the existing one that accepts a single character because it reliably handles line endings.
 * Fixed a bug where the wrong error position was sometimes reported on text with LF line endings.
+* Fixed a bug where the wrong error position was reported on syntax errors.
 
 #### 6.1.0 - 05-03-2021
 * Some typos were fixed in Farkle's HTML templates.